More work on the To Do List
Lots of progress made, we have a lot of console-based functionality now for changing To-Do properties, as well as the Project they are assigned to.

Need to build out the rest of the functionality so that it is fully operational inside the console, then I can work on the DOM manipulation.